api_layer/requirements.txt
google-cloud-bigquery==3.*
functions-framework==3.*
fastapi==0.112.0
uvicorn==0.30.6
flask==3.0.3
gunicorn==23.0.0 